+++ e2e-tests/tests/payment_allocation_fifo_e2e.rs
@@ -44,7 +44,7 @@ async fn create_invoice(
     pool: &sqlx::PgPool,
     tenant_id: &str,
     customer_id: i32,
-    amount_cents: i32,
+    amount_cents: i64,
     currency: &str,
     due_at: Option<&str>,
 ) -> i32 {
